# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: LearnFlow API
  version: 1.0.0
  description: AI-powered learning platform API
servers:
  - url: http://localhost:3000
paths:
...
  /api/v1/update-agent/topics:
    put:
      summary: Update agent topics
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [topics]
              properties:
                topics:
                  type: array
                  items: { type: string }
      responses:
        '200':
          description: Updated
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
  /api/v1/update-agent/sources:
    put:
      summary: Update agent sources
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sources]
              properties:
                sources:
                  type: array
                  items: { type: string }
      responses:
        '200':
          description: Updated
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  headers:
    XRequestId:
      description: Request identifier for tracing
      schema:
        type: string
...